refactor: move styles/ into src/theme/ tree
