+++ src/components/Card/CardCover.tsx
@@ -3,7 +3,7 @@ import { Image, StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
 
 import { getCardCoverStyle } from './utils';
 import { useInternalTheme } from '../../core/theming';
-import { grey200 } from '../../styles/themes/colors';
+import { grey200 } from '../../theme/colors';
 import type { ThemeProp } from '../../types';
 import { splitStyles } from '../../utils/splitStyles';
