+++ src/components/Switch/utils.ts
@@ -9,7 +9,7 @@ import {
   grey700,
   white,
   black,
-} from '../../styles/themes/colors';
+} from '../../theme/colors';
 import type { InternalTheme } from '../../types';
 
 type BaseProps = {
