+++ src/components/TextInput/Adornment/utils.ts
@@ -1,4 +1,4 @@
-import { tokens } from '../../../styles/themes/tokens';
+import { tokens } from '../../../theme/tokens';
 import type { InternalTheme } from '../../../types';
 
 const { stateOpacity } = tokens.md.ref;
