+++ src/components/__tests__/ListAccordion.test.tsx
@@ -4,7 +4,7 @@ import { StyleSheet, View } from 'react-native';
 import { render } from '@testing-library/react-native';
 
 import { getTheme } from '../../core/theming';
-import { red500 } from '../../styles/themes/colors';
+import { red500 } from '../../theme/colors';
 import ListAccordion from '../List/ListAccordion';
 import ListAccordionGroup from '../List/ListAccordionGroup';
 import ListIcon from '../List/ListIcon';
