+++ src/components/__tests__/Snackbar.test.tsx
@@ -3,7 +3,7 @@ import { Animated, StyleSheet, Text, View } from 'react-native';
 
 import { act, render } from '@testing-library/react-native';
 
-import { red200, white } from '../../styles/themes/colors';
+import { red200, white } from '../../theme/colors';
 import Snackbar from '../Snackbar';
 
 const styles = StyleSheet.create({
