+++ src/index.tsx
@@ -1,4 +1,4 @@
-export { Palette } from './styles/themes/tokens';
+export { Palette } from './theme/tokens';
 
 export {
   useTheme,
@@ -8,12 +8,12 @@ export {
   adaptNavigationTheme,
 } from './core/theming';
 
-export * from './styles/themes';
+export * from './theme/schemes';
 
 export { default as Provider } from './core/PaperProvider';
 export { default as PaperProvider } from './core/PaperProvider';
-export { default as shadow } from './styles/shadow';
-export { default as configureFonts } from './styles/fonts';
+export { default as shadow } from './theme/shadow';
+export { default as configureFonts } from './theme/fonts';
 
 import * as Avatar from './components/Avatar/Avatar';
 import * as Drawer from './components/Drawer/Drawer';
