+++ src/theme/__tests__/fonts.test.js
@@ -130,7 +130,7 @@ const loadFonts = () => {
     const fonts = require('../fonts');
     configureFonts = fonts.default;
 
-    typescale = require('../themes/tokens').typescale;
+    typescale = require('../tokens').typescale;
   });
 
   return { configureFonts, typescale };
