+++ src/theme/types.ts
@@ -1,0 +1,179 @@
+import type * as React from 'react';
+
+import type { $DeepPartial } from '@callstack/react-theme-provider';
+
+export type Font = {
+  fontFamily: string;
+  fontWeight?:
+    | 'normal'
+    | 'bold'
+    | '100'
+    | '200'
+    | '300'
+    | '400'
+    | '500'
+    | '600'
+    | '700'
+    | '800'
+    | '900';
+  fontStyle?: 'normal' | 'italic' | undefined;
+};
+
+export type Fonts = {
+  regular: Font;
+  medium: Font;
+  light: Font;
+  thin: Font;
+};
+
+type Mode = 'adaptive' | 'exact';
+
+export type ThemeColors = {
+  primary: string;
+  primaryContainer: string;
+  secondary: string;
+  secondaryContainer: string;
+  tertiary: string;
+  tertiaryContainer: string;
+  surface: string;
+  surfaceDim: string;
+  surfaceBright: string;
+  surfaceContainerLowest: string;
+  surfaceContainerLow: string;
+  surfaceContainer: string;
+  surfaceContainerHigh: string;
+  surfaceContainerHighest: string;
+  surfaceVariant: string;
+  background: string;
+  error: string;
+  errorContainer: string;
+  onPrimary: string;
+  onPrimaryContainer: string;
+  onSecondary: string;
+  onSecondaryContainer: string;
+  onTertiary: string;
+  onTertiaryContainer: string;
+  onSurface: string;
+  onSurfaceVariant: string;
+  onError: string;
+  onErrorContainer: string;
+  onBackground: string;
+  outline: string;
+  outlineVariant: string;
+  inverseSurface: string;
+  inverseOnSurface: string;
+  inversePrimary: string;
+  primaryFixed: string;
+  primaryFixedDim: string;
+  onPrimaryFixed: string;
+  onPrimaryFixedVariant: string;
+  secondaryFixed: string;
+  secondaryFixedDim: string;
+  onSecondaryFixed: string;
+  onSecondaryFixedVariant: string;
+  tertiaryFixed: string;
+  tertiaryFixedDim: string;
+  onTertiaryFixed: string;
+  onTertiaryFixedVariant: string;
+  shadow: string;
+  scrim: string;
+  /** Pre-computed state layer color at press opacity (0.10).
+   *  Used for ripple effects. Avoids runtime alpha manipulation
+   *  which is incompatible with PlatformColor on Android.
+   *  TODO: revisit after https://github.com/facebook/react-native/pull/56395
+   *  @see https://m3.material.io/foundations/interaction/states/state-layers */
+  stateLayerPressed: string;
+  elevation: ElevationColors;
+};
+
+export type ThemeProp = $DeepPartial<InternalTheme>;
+
+export type ThemeBase = {
+  dark: boolean;
+  mode?: Mode;
+  roundness: number;
+  animation: {
+    scale: number;
+    defaultAnimationDuration?: number;
+  };
+};
+
+export type Theme = ThemeBase & {
+  colors: ThemeColors;
+  fonts: Typescale;
+};
+
+export type InternalTheme = Theme;
+
+export enum TypescaleKey {
+  displayLarge = 'displayLarge',
+  displayMedium = 'displayMedium',
+  displaySmall = 'displaySmall',
+
+  headlineLarge = 'headlineLarge',
+  headlineMedium = 'headlineMedium',
+  headlineSmall = 'headlineSmall',
+
+  titleLarge = 'titleLarge',
+  titleMedium = 'titleMedium',
+  titleSmall = 'titleSmall',
+
+  labelLarge = 'labelLarge',
+  labelMedium = 'labelMedium',
+  labelSmall = 'labelSmall',
+
+  bodyLarge = 'bodyLarge',
+  bodyMedium = 'bodyMedium',
+  bodySmall = 'bodySmall',
+}
+
+export type TypescaleStyle = {
+  fontFamily: string;
+  letterSpacing: number;
+  fontWeight: Font['fontWeight'];
+  lineHeight: number;
+  fontSize: number;
+  fontStyle?: Font['fontStyle'];
+};
+
+export type Typescale =
+  | {
+      [key in TypescaleKey]: TypescaleStyle;
+    } & {
+      ['default']: Omit<TypescaleStyle, 'lineHeight' | 'fontSize'>;
+    };
+
+export type Elevation = 0 | 1 | 2 | 3 | 4 | 5;
+
+export enum ElevationLevels {
+  'level0',
+  'level1',
+  'level2',
+  'level3',
+  'level4',
+  'level5',
+}
+
+export type ElevationColors = {
+  [key in keyof typeof ElevationLevels]: string;
+};
+
+export type $Omit<T, K> = Pick<T, Exclude<keyof T, K>>;
+export type $RemoveChildren<T extends React.ComponentType<any>> = $Omit<
+  React.ComponentPropsWithoutRef<T>,
+  'children'
+>;
+
+export type EllipsizeProp = 'head' | 'middle' | 'tail' | 'clip';
+
+export type NavigationTheme = {
+  dark: boolean;
+  colors: {
+    primary: string;
+    background: string;
+    card: string;
+    text: string;
+    border: string;
+    notification: string;
+  };
+};
